% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GET}
\alias{GET}
\title{GET request data from Spotify with limit and offset parameters}
\usage{
GET(url, params, oAuth = get_access_token())
}
\arguments{
\item{url}{a Spotify URL request}

\item{params}{a list that contains limit and offset values}

\item{oAuth}{A valid access token from the Spotify Accounts service.}
}
\value{

}
\description{
GET request data from Spotify with limit and offset parameters
}
