% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_access_token}
\alias{get_access_token}
\title{Creates an access token to access Spotify web API.}
\usage{
get_access_token(
  client_id = Sys.getenv("SPOTIFY_CLIENT_ID"),
  client_secret = Sys.getenv("SPOTIFY_CLIENT_SECRET"),
  scope = spotifyr::scopes
)
}
\arguments{
\item{client_id}{System Environment variable "SPOTIFY_CLIENT_ID"}

\item{client_secret}{System Environment variable "SPOTIFY_CLIENT_SECRET"}

\item{scope}{Space delimited string of spotify scopes. All scopes are
selected by default}
}
\description{
Creates a long-lived access token that allows R to make calls to
  the Spotify API.
}
