% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{get_artist_albums}
\alias{get_artist_albums}
\title{Get Spotify catalog information about an artist’s albums.}
\usage{
get_artist_albums(artist.id, limit = 50, offset = 0, country = "")
}
\arguments{
\item{artist.id}{The Spotify ID for the artist.}

\item{limit}{Maximum number of results to return}

\item{offset}{The index of the first result to return. Use with limit to get
the next page of search results.}

\item{country}{An ISO 3166-1 alpha-2 country code. See details
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO 3166-1 alpha-2}}
}
\value{
Spotify catalog information about an artist’s albums
}
\description{
Get Spotify catalog information about an artist’s albums.
}
