% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{get_categories}
\alias{get_categories}
\title{Get a List of Categories}
\usage{
get_categories(limit = 50, offset = 0)
}
\arguments{
\item{limit}{Maximum number of results to return}

\item{offset}{The index of the first result to return. Use with limit to get
the next page of search results.}
}
\value{
a List of Categories available on Spotify
}
\description{
Get a List of Categories
}
