% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{get_categories_playlists}
\alias{get_categories_playlists}
\title{Get a list of Spotify playlists tagged with a particular category.}
\usage{
get_categories_playlists(category.id, limit = 50, offset = 0)
}
\arguments{
\item{category.id}{The Spotify category ID for the category.}

\item{limit}{Maximum number of results to return}

\item{offset}{The index of the first result to return. Use with limit to get
the next page of search results.}
}
\value{
a list of Spotify category's playlists
}
\description{
Get a list of Spotify playlists tagged with a particular category.
}
