% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{get_multiple_audio_features}
\alias{get_multiple_audio_features}
\title{Get audio features for multiple tracks based on their Spotify IDs.}
\usage{
get_multiple_audio_features(track.ids = c(), limit = 100, offset = 0)
}
\arguments{
\item{track.ids}{a string vector of the Spotify track IDs}

\item{limit}{Maximum number of results to return}

\item{offset}{The index of the first result to return. Use with limit to get
the next page of search results.}
}
\value{
a list of audio features of the multiple tracks
}
\description{
Get audio features for multiple tracks based on their Spotify IDs.
}
