% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{get_multiple_tracks_info}
\alias{get_multiple_tracks_info}
\title{Get Spotify catalog information for multiple tracks based on their Spotify IDs.}
\usage{
get_multiple_tracks_info(track.ids, limit = 50, offset = 0)
}
\arguments{
\item{track.ids}{a string vector of the Spotify track IDs}

\item{limit}{Maximum number of results to return}

\item{offset}{The index of the first result to return. Use with limit to get
the next page of search results.}
}
\value{
a list of Spotify catalog information for multiple tracks
}
\description{
Get Spotify catalog information for multiple tracks based on their Spotify IDs.
}
