% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{get_new_release}
\alias{get_new_release}
\title{Get a list of new album releases featured in Spotify}
\usage{
get_new_release(limit = 50, offset = 0)
}
\arguments{
\item{limit}{Maximum number of results to return}

\item{offset}{The index of the first result to return. Use with limit to get
the next page of search results.}
}
\value{
a list of new album release
}
\description{
Get a list of new album releases featured in Spotify
}
