% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{get_one_audio_analysis}
\alias{get_one_audio_analysis}
\title{Get a detailed audio analysis for a single track identified by its unique
Spotify ID.}
\usage{
get_one_audio_analysis(track.id)
}
\arguments{
\item{track.id}{a string of one Spotify track's ID}
}
\value{
a detailed audio analysis for a single track. See details
  \href{https://developer.spotify.com/documentation/web-api/reference/tracks/get-audio-analysis/}{Spotify
   Web API}
}
\description{
Get a detailed audio analysis for a single track identified by its unique
Spotify ID.
}
