% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{get_one_playlist_info}
\alias{get_one_playlist_info}
\title{Get a playlist information}
\usage{
get_one_playlist_info(playlist.id, limit = 50, offset = 0, filter = TRUE)
}
\arguments{
\item{playlist.id}{The Spotify ID for the playlist}

\item{limit}{Maximum number of results to return}

\item{offset}{The index of the first result to return. Use with limit to get
the next page of search results.}

\item{filter}{a boolean to return all data or filtered data}
}
\value{
a list of information regarding to a playlist
}
\description{
Get a playlist information
}
