% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{get_one_tracks_info}
\alias{get_one_tracks_info}
\title{Get Spotify catalog information for a single track identified by its unique Spotify ID.}
\usage{
get_one_tracks_info(track.id, limit = 50, offset = 0)
}
\arguments{
\item{track.id}{a string of one Spotify track's ID}

\item{limit}{Maximum number of results to return}

\item{offset}{The index of the first result to return. Use with limit to get
the next page of search results.}
}
\value{
Spotify catalog informatuon for a single track
}
\description{
Get Spotify catalog information for a single track identified by its unique Spotify ID.
}
