% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{get_playlist_byUser}
\alias{get_playlist_byUser}
\title{Get a list of the playlists owned or followed by a Spotify user.}
\usage{
get_playlist_byUser(user.id, limit = 50, offset = 0, filter = TRUE)
}
\arguments{
\item{user.id}{The Spotify user ID}

\item{limit}{Maximum number of results to return}

\item{offset}{The index of the first result to return. Use with limit to get
the next page of search results.}

\item{filter}{a boolean to return all data or filtered data}
}
\value{
a list of playlists
}
\description{
Get a list of the playlists owned or followed by a Spotify user.
}
