% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{get_related_artists}
\alias{get_related_artists}
\title{Get Spotify catalog information about artists similar to a given artist.}
\usage{
get_related_artists(artist.id, limit = 50, offset = 0)
}
\arguments{
\item{artist.id}{The Spotify ID for the artist.}

\item{limit}{Maximum number of results to return}

\item{offset}{The index of the first result to return. Use with limit to get
the next page of search results.}
}
\value{
a list of Spotify catalog information about artists similar to a
  given artist.
}
\description{
Get Spotify catalog information about artists similar to a given artist.
}
