% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users_profile.R
\name{get_user_info}
\alias{get_user_info}
\title{Get public profile information about a Spotify user.}
\usage{
get_user_info(user.id = "", limit = 50, offset = 0)
}
\arguments{
\item{user.id}{The Spotify user ID}

\item{limit}{Maximum number of results to return}

\item{offset}{The index of the first result to return. Use with limit to get
the next page of search results.}
}
\value{
a list of public profile information about a Spotify user. See details
  \href{https://developer.spotify.com/documentation/web-api/reference/users-profile/get-users-profile/}{Spotify Web API}
}
\description{
Get public profile information about a Spotify user.
}
