% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_artist_byName}
\alias{search_artist_byName}
\title{Get Spotify Catalog information about artists}
\usage{
search_artist_byName(
  search_word,
  limit = 50,
  offset = 0,
  include_external = audio
)
}
\arguments{
\item{search_word}{search query keyword}

\item{limit}{Maximum number of results to return}

\item{offset}{The index of the first result to return. Use with limit to get
the next page of search results.}

\item{include_external}{include_external=audio is specified the response will
include any relevant audio content that is hosted externally.}
}
\value{
a list of Spotify Catalog information about artists
}
\description{
Get Spotify Catalog information about artists
}
