% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotifyrLite-package.R
\docType{package}
\name{spotifyrLite-package}
\alias{spotifyrLite}
\alias{spotifyrLite-package}
\title{A lightweight R wrapper to access Spotify API}
\description{
This package allows you to make requests to Spotify Web API to
  retrieve information such as audio features, audio analysis for a track,
  artist's top tracks, similar artists given a artist, etc.
  Although the package does not included all the available requests to
  Spotify's API, important requests such as search for an item in Spotify are
  included to allow you to retrieve playlists, albums, artists or tracks of
  interest. This subsequently allows you to collect audio features or audio
  analysis of the respective tracks within the playlists or albums retrieved
  from the previous search.
}
\details{
See the README on \href{https://github.com/Giant316/spotifyLite/blob/main/README.md}{Github}
}
\keyword{internal}
