% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xGET}
\alias{xGET}
\title{GET request data from Spotify without limit and offset parameters}
\usage{
xGET(url, oAuth = get_access_token())
}
\arguments{
\item{url}{a Spotify URL request}

\item{oAuth}{A valid access token from the Spotify Accounts service.}
}
\value{

}
\description{
GET request data from Spotify without limit and offset parameters
}
